Reorganizar archivos Java en carpetas
Se agruparon los archivos Java en carpetas correspondientes para mejorar la estructura del proyecto y la facilidad de navegación. Cada grupo de archivos relacionados ahora reside en su propia carpeta, reflejando mejor la modularidad del sistema.
