+++ src/main/java/org/finder/FileReaders/DocxFileReader.java
@@ -1,9 +1,11 @@
-package org.finder;
+package org.finder.FileReaders;
 
 import org.apache.poi.xwpf.usermodel.XWPFDocument;
 import org.apache.poi.xwpf.usermodel.XWPFParagraph;
 import org.apache.logging.log4j.LogManager;
 import org.apache.logging.log4j.Logger;
+import org.finder.Tree.AVLTree;
+
 import java.io.FileInputStream;
 import java.io.IOException;
 import java.util.List;
