+++ src/main/java/org/finder/FileReaders/TextFileReader.java
@@ -1,9 +1,11 @@
-package org.finder;
+package org.finder.FileReaders;
 import java.io.BufferedReader;
 import java.io.FileReader;
 import java.io.IOException;
 import org.apache.logging.log4j.LogManager;
 import org.apache.logging.log4j.Logger;
+import org.finder.Tree.AVLTree;
+
 /**
  * Clase encargada de leer archivos de texto y extraer palabras para insertarlas en un árbol AVL.
  * Esta clase ofrece la funcionalidad de procesar archivos de texto, extrayendo cada palabra y
